+++ package.json
@@ -14,7 +14,9 @@
         "@aws-sdk/client-dynamodb": "^3.758.0",
         "@aws-sdk/lib-dynamodb": "^3.758.0",
         "env-var": "^7.5.0",
-        "gramio": "^0.2.4"
+        "gramio": "^0.2.4",
+        "pino": "^9.6.0",
+        "pino-pretty": "^13.0.0"
     },
     "description": "Telegram bot that forwards messages between users and a group with topics",
     "devDependencies": {
