+++ src/middlewares/withBot.ts
@@ -4,10 +4,6 @@ export const withBot = async (ctx: ForwardContext, next: () => Promise<void>) =>
     const me = await ctx.api?.getMe();
     ctx.me = me;
 
-    return next();
-};
-
-export const skipSelfMessages = async (ctx: ForwardContext, next: () => Promise<void>) => {
     if (ctx.from?.id !== ctx.me?.id) {
         return next();
     }
