+++ src/middlewares/withDB.ts
@@ -1,6 +1,9 @@
 import type { DynamoDBService } from '@/services/dynamodb.js';
+import type { ForwardContext } from '@/types.js';
 
-export const withDB = (db: DynamoDBService) => async (ctx: any, next: () => Promise<void>) => {
-    ctx.db = db;
-    return await next();
+export const withDB = (db: DynamoDBService) => {
+    return (ctx: ForwardContext, next: () => Promise<void>) => {
+        ctx.db = db;
+        return next();
+    };
 };
